+++ src/lib/src/theme/common.module.ts
@@ -2,10 +2,20 @@ import { NgModule } from '@angular/core';
 
 import { LyPaper } from './paper';
 import { LyWithClass } from './with-class.directive';
-import { LyStyle } from './style.directive';
+import { LyStyle, LyStyleDeprecated } from './style.directive';
 
 @NgModule({
   declarations: [LyStyle, LyWithClass, LyPaper],
   exports: [LyStyle, LyWithClass, LyPaper]
 })
 export class LyCommonModule { }
+
+/**
+ * @deprecated
+ * LyStyle without prefixes
+ */
+@NgModule({
+  declarations: [LyStyleDeprecated],
+  exports: [LyStyleDeprecated]
+})
+export class LyCommonDeprecatedModule { }